SAGE-752: use server:port from reverse-tunnel config
The reverse tunnel server and port number are not read from the
system config and used to construct the pattern to search for within
the "ssh connection lookup" ('ss'). The pattern is constructed of
the host's IP address and port number ('<ip>:<port>') to eliminate the
need for 'ss' to perform any network resolution.