SAGE-1414: add pass/fail history system
Before this change a single connection pass reset the recovery score-board.
This presented problems with "blippy" internet that required a hard-reboot.
Because the internet test would pass 1 time during an entire boot, the
recovery score-board would be reset and a hard-reboot would never be
attempted.

This change adds a connection history buffer requiring a pass/fail
majority before a connection is considered good/bad again. The last X
number of test results are averaged to "connection confidence"
percentage. When that percentage exceeds a "healthy" threshold the
connection is considered healthy and the scoreboard is reset. When
the percentage falls below a threshold, the connection is considered
unheathly, and recovery actions are initiated.

Other notable changes:
- remove the 3x "successive" tests per run (replaced by this sytem)
- removed 1 soft reboot
- reduced frequency of network restarts (which tend not to help)
- added connection history buffer
- updated unit-test to have passing and failure connection checks