SAGE-1414: add "pass window" support
Before this change a single connection pass reset the recovery score-board.
This presented problems with "blippy" internet that required a hard-reboot.
Because the internet test would pass 1 time during an entire boot, the
recovery score-board would be reset and a hard-reboot would never be
attempted.

This change requires that a window of consecutive passes occur in-order
to reset the recovery score-board.  This way a single passing connection
test or very short passing window does not cause the recovery process
to be interrupted.

Other notable changes:
- removed 1 soft reboot
- reduced frequency of network restarts (which tend not to help)
- added 10 minute "pass window" before score-board reset
- updated unit-test to have passing and failure connection checks