SAGE-1414: add pass/fail buffer system
Before this change a single connection pass reset the recovery score-board.
This presented problems with "blippy" internet that required a hard-reboot.
Because the internet test would pass 1 time during an entire boot, the
recovery score-board would be reset and a hard-reboot would never be
attempted.

This change adds a buffer requiring a pass/fail majority before a
connection is considered good/bad again. In order for a connection
to be considered good, there needs to be a long majority of passing
connection tests to push the "connection confidence" high enough
for a connection to be considered passing. The inverse applies for
detection of a bad connection, driving the "connection confidence" low.

Other notable changes:
- remove the 3x "successive" tests per run (replaced by this sytem)
- removed 1 soft reboot
- reduced frequency of network restarts (which tend not to help)
- added confidence buffer
- updated unit-test to have passing and failure connection checks