SAGE-1414: WIP: we are pausing on this strategy as this too easily enters recovery
BASICALLY WITH A LARGE REQUIRED PASS WINDOW, WE ARE AT RISK OF RANDOM
FAILURES CAUSING THE SYSTEM TO ENTER QUICK RECOVERY

basically since we require a large window for success, and now a single
"failure blip" resets that success window, we can have over a 10 minute
period, 2 random blips, cause for the 10 minute success window to be
reset easily and the system to enter recovery.

going to try a new mechanism of a [0 - max], when we hit max, we reset
the scoreboard, and when we hit 0, we enter a recovery step.

essentially, we build up confidence and it requires an overweight of
failures to cause the system to degrade